reorganized graphql files into src/graphql
