+++ .gitignore
@@ -9,7 +9,7 @@
 /src/locales/**/pseudo.po
 
 # generated graphql types
-/src/schema
+/src/graphql/schema/
 __generated__/
 
 # dependencies
