+++ codegen.yml
@@ -2,6 +2,6 @@ overrideExisting: true
 schema: 'https://api.thegraph.com/subgraphs/name/uniswap/uniswap-v3'
 documents: 'src/**/!(*.d).{ts,tsx}'
 generates:
-  ./src/schema/schema.graphql:
+  ./src/graphql/schema/schema.graphql:
     plugins:
       - schema-ast
