+++ src/graphql/AllV3TicksQuery.ts
