+++ src/index.tsx
@@ -4,6 +4,7 @@ import 'polyfills'
 import 'components/analytics'
 
 import { FeatureFlagsProvider } from 'featureFlags'
+import RelayEnvironment from 'graphql/RelayEnvironment'
 import { BlockNumberProvider } from 'lib/hooks/useBlockNumber'
 import { MulticallUpdater } from 'lib/state/multicall'
 import { StrictMode } from 'react'
@@ -12,7 +13,6 @@ import { QueryClient, QueryClientProvider } from 'react-query'
 import { Provider } from 'react-redux'
 import { HashRouter } from 'react-router-dom'
 import { RelayEnvironmentProvider } from 'relay-hooks'
-import RelayEnvironment from 'RelayEnvironment'
 
 import Blocklist from './components/Blocklist'
 import Web3Provider from './components/Web3Provider'
