+++ src/utils/computeSurroundingTicks.ts
@@ -1,8 +1,9 @@
 import { Token } from '@uniswap/sdk-core'
 import { tickToPrice } from '@uniswap/v3-sdk'
-import { Ticks } from 'hooks/graphql/AllV3TicksQuery'
 import { TickProcessed } from 'hooks/usePoolTickData'
 import JSBI from 'jsbi'
+
+import { Ticks } from '../graphql/AllV3TicksQuery'
 
 const PRICE_FIXED_DIGITS = 8
